refactor: 내 작품 모아보기 페이지 Server Component 전환
