+++ src/app/artworks/page.tsx
@@ -1,10 +1,7 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-import ArtworkCard from '@/components/myArtworks/ArtworkCard';
-import ArtworkModal from '@/components/myArtworks/ArtworkModal';
-import FilterTab from '@/components/myArtworks/FilterTab';
-import { Artwork, FilterType } from '@/components/myArtworks/Types';
+import { redirect } from 'next/navigation';
+import { createClient } from '@/lib/supabase/server';
+import type { Artwork } from '@/components/myArtworks/Types';
+import ArtworksScreen from '@/components/myArtworks/ArtworksScreen';
 
 type RawArtwork = {
   id: string;
@@ -13,129 +10,46 @@ type RawArtwork = {
   description: string | null;
   image_url: string | null;
   created_at: string;
-  artwork_likes: { count: number }[];
+  artwork_likes: { count: number }[] | null;
   exhibitions: {
     title: string;
     profiles: { institution: string | null } | null;
   } | null;
 };
 
-function mapArtwork(raw: RawArtwork): Artwork {
-  return {
-    id: raw.id,
-    title: raw.title,
-    artist: raw.artist_name ?? '',
-    description: raw.description ?? '',
-    exhibitionTitle: raw.exhibitions?.title ?? '',
-    academyName: raw.exhibitions?.profiles?.institution ?? '',
-    imageUrl: raw.image_url ?? '',
-    likesCount: (raw.artwork_likes ?? [])[0]?.count ?? 0,
-    createdAt: raw.created_at,
-  };
-}
+export default async function MyArtworksPage() {
+  const supabase = await createClient();
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
 
-export default function MyArtworksPage() {
-  const [artworks, setArtworks] = useState<Artwork[]>([]);
-  const [filter, setFilter] = useState<FilterType>('latest');
-  const [search, setSearch] = useState<string>('');
-  const [selectedArtwork, setSelectedArtwork] = useState<Artwork | null>(null);
+  if (!user) redirect('/login');
 
-  useEffect(() => {
-    fetch('/api/archive?type=mine')
-      .then((res) => res.json())
-      .then((json) => setArtworks((json.artworks ?? []).map(mapArtwork)))
-      .catch(console.error);
-  }, []);
+  const { data, error } = await supabase
+    .from('artworks')
+    .select(
+      `id, title, artist_name, description, image_url, created_at,
+      artwork_likes(count),
+      exhibitions ( title, profiles!teacher_id ( institution ) )`
+    )
+    .eq('artist_id', user.id)
+    .order('created_at', { ascending: false });
 
-  const sorted = [...artworks]
-    .filter((a) => a.title.toLowerCase().includes(search.toLowerCase()))
-    .sort((a, b) => {
-      if (filter === 'latest')
-        return (
-          new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-        );
-      if (filter === 'oldest')
-        return (
-          new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime()
-        );
-      return b.likesCount - a.likesCount;
-    });
+  if (error) throw new Error(error.message);
 
-  return (
-    <>
-      <main className="min-h-screen bg-[#F5F0E8]">
-        <div className="mx-auto max-w-[1080px] px-6 py-10">
-          <div className="mb-6 flex items-start justify-between">
-            <div>
-              <h1 className="mb-1 text-[28px] font-bold tracking-tight text-[#1A1A1A]">
-                내 작품 모아보기
-              </h1>
-              <p className="text-[14px] text-[#888780]">
-                나의 작품 {artworks.length}점
-              </p>
-            </div>
-
-            <div className="relative">
-              <svg
-                className="absolute top-1/2 left-3.5 -translate-y-1/2 text-[#BCBAB2]"
-                width="15"
-                height="15"
-                viewBox="0 0 16 16"
-                fill="none"
-              >
-                <circle
-                  cx="7"
-                  cy="7"
-                  r="5.5"
-                  stroke="currentColor"
-                  strokeWidth="1.5"
-                />
-                <path
-                  d="M11 11l2.5 2.5"
-                  stroke="currentColor"
-                  strokeWidth="1.5"
-                  strokeLinecap="round"
-                />
-              </svg>
-              <input
-                type="text"
-                placeholder="작품 검색..."
-                value={search}
-                onChange={(e) => setSearch(e.target.value)}
-                className="w-[220px] rounded-full border border-[#EDEBE4] bg-white py-2.5 pr-4 pl-9 text-[14px] text-[#1A1A1A] transition-colors outline-none placeholder:text-[#BCBAB2] focus:border-[#f4b942]"
-              />
-            </div>
-          </div>
-
-          <FilterTab value={filter} onChange={setFilter} />
-
-          {sorted.length > 0 ? (
-            <div className="grid grid-cols-4 gap-5">
-              {sorted.map((artwork) => (
-                <ArtworkCard
-                  key={artwork.id}
-                  artwork={artwork}
-                  onClick={() => setSelectedArtwork(artwork)}
-                />
-              ))}
-            </div>
-          ) : (
-            <div className="flex flex-col items-center justify-center py-32 text-[#BCBAB2]">
-              <span className="mb-4 text-5xl">🎨</span>
-              <p className="text-[15px] font-medium text-[#888780]">
-                검색 결과가 없습니다.
-              </p>
-            </div>
-          )}
-        </div>
-      </main>
-
-      {selectedArtwork && (
-        <ArtworkModal
-          artwork={selectedArtwork}
-          onClose={() => setSelectedArtwork(null)}
-        />
-      )}
-    </>
+  const artworks: Artwork[] = ((data ?? []) as unknown as RawArtwork[]).map(
+    (raw) => ({
+      id: raw.id,
+      title: raw.title,
+      artist: raw.artist_name ?? '',
+      description: raw.description ?? '',
+      exhibitionTitle: raw.exhibitions?.title ?? '',
+      academyName: raw.exhibitions?.profiles?.institution ?? '',
+      imageUrl: raw.image_url ?? '',
+      likesCount: (raw.artwork_likes ?? [])[0]?.count ?? 0,
+      createdAt: raw.created_at,
+    })
   );
+
+  return <ArtworksScreen artworks={artworks} />;
 }
